requirements.txt
fastapi
uvicorn
pymongo==4.6.1
dnspython==2.8.0

python-dotenv
passlib[bcrypt]
python-jose
pydantic[email]
